REFACTOR traits & interfaces out of encapsulated namespaces
