+++ tests/Feature/Builders/TrackActionBuilderTest.php
@@ -3,7 +3,6 @@
 namespace Sfneal\Tracking\Tests\Feature\Builders;
 
 use Sfneal\Tracking\Models\TrackAction;
-use Sfneal\Tracking\Tests\Feature\Builders\Traits\WhereModelTests;
 
 class TrackActionBuilderTest extends BuilderTestCase
 {
