+++ tests/Feature/Builders/TrackActivityBuilderTest.php
@@ -3,8 +3,6 @@
 namespace Sfneal\Tracking\Tests\Feature\Builders;
 
 use Sfneal\Tracking\Models\TrackActivity;
-use Sfneal\Tracking\Tests\Feature\Builders\Traits\WhereModelTests;
-use Sfneal\Tracking\Tests\Feature\Builders\Traits\WhereUserTests;
 
 class TrackActivityBuilderTest extends BuilderTestCase
 {
