+++ tests/Feature/Builders/TrackTrafficBuilderTest.php
@@ -3,7 +3,6 @@
 namespace Sfneal\Tracking\Tests\Feature\Builders;
 
 use Sfneal\Tracking\Models\TrackTraffic;
-use Sfneal\Tracking\Tests\Feature\Builders\Traits\WhereUserTests;
 
 class TrackTrafficBuilderTest extends BuilderTestCase
 {
