+++ tests/Feature/Factories/TrackActionFactoryTest.php
@@ -3,7 +3,6 @@
 namespace Sfneal\Tracking\Tests\Feature\Factories;
 
 use Sfneal\Tracking\Models\TrackAction;
-use Sfneal\Tracking\Tests\Feature\Factories\Interfaces\FactoryFillablesTest;
 
 class TrackActionFactoryTest extends FactoriesTestCase implements FactoryFillablesTest
 {
