+++ tests/Feature/Factories/TrackActivityFactoryTest.php
@@ -3,7 +3,6 @@
 namespace Sfneal\Tracking\Tests\Feature\Factories;
 
 use Sfneal\Tracking\Models\TrackActivity;
-use Sfneal\Tracking\Tests\Feature\Factories\Interfaces\FactoryFillablesTest;
 
 class TrackActivityFactoryTest extends FactoriesTestCase implements FactoryFillablesTest
 {
